requirements.txt
langchain-azure-ai
flask==2.3.3
python-dotenv==1.0.0
gunicorn==21.2.0
uuid==1.30